chore: use the sensible build... compatibly
We try to use the "sensible" build when possible, handling all the old
try/catch magic in catsrc. This mostly works, except for a problem in
the parser that appears to be also present in my previous version: a
huge object tries to reference things that don't exist.

[The site is unchanged for now to avoid breaking any new things.
Somebody added a // comment to CSS which is invalid, so expect it to not
work. I did not break it.]

The eslint rules are enabled. They don't really hit any package problems
any more, so I figure it's time to turn them on. I added an rc for the
tests so every complaint left is a true positive now.

The Makefile has been rewritten to be aware of new changes. In the long
run webpack would be the way to go still.