stdlib: Improve sin, cos and tan (^/1/3)
Robust: ^/4 (Does it work for all input values?)
Accurate: 1/4 (How accurate is the result?)
Well behaved: 3/4 (Does it preserve math properties?)

^ They technically work for all input values. But for very large input,
the loss of accuracy makes them practically useless.

The result could be severely inaccurate either when the angle is very
large, or when the angle is very close to a multiple of pi/2.